+++ components/Button.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import * as React from 'react'
 
 import { Button as ButtonPrimitive } from 'react-aria-components'
