+++ styles/index.css
@@ -84,4 +84,10 @@
   #__next {
     isolation: isolate;
   }
+
+  /* Shiki */
+  .shiki {
+    background-color: transparent !important;
+    color: inherit !important;
+  }
 }
